test: split integration tests into multiple packages
